Functions

This page lists all Modules, Functions and Structures available in this package.

QGDipoles.jl

QGDipoles — Module

Package for creating steady modon solutions to the layered quasi-geostrophic equations and the surface quasi geostrophic equations.

See examples/ for example scripts and here for documentation.

JJ_integ.jl

QGDipoles.A_func — Function

Function: A_func(ξ, λ, μ)

Evaluates the matrix function A(ξ, λ, μ) = K(ξ) [K(ξ) + D(μ)]⁻¹ ξ⁻¹

Arguments:

ξ: point in [0, ∞), Number

λ: ratio of vortex radius to Rossby radius in each layer, Number or Vector

μ: nondimensional (y) vorticity gradient in each layer, Number or Vector

QGDipoles.B_func — Function

Function: B_func(ξ, λ, μ)

Evaluates the matrix function B(ξ, λ, μ) = [K(ξ) + D(μ)]⁻¹ ξ⁻¹

Arguments:

ξ: point in [0, ∞), Number

λ: ratio of vortex radius to Rossby radius in each layer, Number or Vector

μ: nondimensional (y) vorticity gradient in each layer, Number or Vector

QGDipoles.JJ_int — Function

Function: JJ_int(F, j, k, tol=1e-6)

Evaluates the integral $I = \int_0^\infty F(\xi) J_{2j+2}(\xi) J_{2k+2}(\xi) \mathrm{d}\xi$

Arguments:

F: function to integrate, typically A_func or B_func, Function

j: first Bessel function index, Integer

k: second Bessel function index, Integer

tol: error tolerance for QuadGK, Number (default: 1e-6)

Note: This integral is performed by deforming the contour of integration into the complex plane where the Bessel function decays exponentially in the imaginary direction.

lin_sys.jl

QGDipoles.BuildLinSys — Function

Function: BuildLinSys(M, λ, μ; tol=1e-6, sqg=false)

Builds the terms in the inhomogeneous eigenvalue problem; A, B, c, d

Arguments:

M: number of coefficient to solve for, Integer

λ: ratio of vortex radius to Rossby radius in each layer, Number or Vector

μ: nondimensional (y) vorticity gradient in each layer, Number or Vector

tol: error tolerance for QuadGK via JJ_int, Number (default: 1e-6)

sqg: false; creates layered QG system, true; creates SQG system (default: false)

QGDipoles.ApplyPassiveLayers — Function

Function: ApplyPassiveLayers(A, B, c, d, ActiveLayers)

Removes rows and columns corresponding to passive layers from the system

Arguments:

A, B, c, d: inhomogeneous eigenvalue problem terms, Arrays

ActiveLayers: vector of 1s or 0s where 1 denotes an active layer, Number or Vector

QGDipoles.IncludePassiveLayers — Function

Function: IncludePassiveLayers(K, a, ActiveLayers)

Includes columns corresponding to passive layers in the eigenvalue and coefficient arrays

Arguments:

K, a: eigenvalue and coefficient arrays describing system solution, Arrays

ActiveLayers: vector of 1s or 0s where 1 denotes an active layer, Number or Vector

QGDipoles.SolveInhomEVP — Function

Function: SolveInhomEVP(A, B, c, d; K₀=Nothing, a₀=Nothing, tol=1e-6, method=0, m=2, sqg=false)

Solves the inhomogeneous eigenvalue problem using nonlinear root finding

Arguments:

A, B, c, d: inhomogeneous eigenvalue problem terms, Arrays

K₀, a₀: initial guesses for K and a, Arrays or Nothings (default: Nothing)

tol: error tolerance for nlsolve, Number (default: 1e-6)

method: 0 - eigensolve for N = 1 and nlsolve for N > 1, 1 - nlsolve (default: 0)

m: exponent of K in eignevalue problem (default: 2)

sqg: false, uses m value specified; true, sets m=1 (default: false)

Note: setting sqg=true overwrites the value of m and is equivalent to setting m=1. The option to set both is included for consistency with BuildLinSys and more generality with the value of m.

QGDipoles.InhomEVP_F! — Function

Function: InhomEVP_F!(F, J, x, A, B, c, d, e)

Calculates the function F and it's derivatives, J, at a given point x

Arguments:

F, J: values of F and it's derivatives, updated by function

x: evaluation point, Array

A, B, c: inhomogeneous eigenvalue problem terms, Arrays

e: basis spanning the space perpendicular to the d[n], Array

QGDipoles.OrthogSpace — Function

Function: OrthogSpace(v)

Extends the input to an orthonormal basis over R^n using the Gram-Schmidt method

Arguments:

v: array with vectors as columns, Array

create_modon.jl

QGDipoles.ZernikeR — Function

Function: ZernikeR(n, x)

Define the Zernike radial function using the jacobi function from SpecialFunctions

Arguments:

n: order, Integer

x: evaluation point, Number or Array

QGDipoles.GridStruct — Type

Structure: GridStruct

Stores the grid variables in physical and Fourier space

Arguments:

x, y: x and y points in physical space, Ranges

kr, l: x and y points in Fourier space, Arrays

Krsq: kr²+l² in Fourier space, Array

QGDipoles.CreateGrid — Function

Function: CreateGrid(Nx, Ny, Lx, Ly; cuda=false)

Define the numerical grid as a GridStruct

Arguments:

Nx, Ny: number of gridpoints in x and y directions, Integers

Lx, Ly: x and y domains, either vectors of endpoints or lengths, Vectors or Numbers

cuda: true; use CUDA CuArray for fields (default: false)

QGDipoles.Calc_ψq — Function

Function: Calc_ψq(a, U, ℓ, R, β, grid, x₀=[0, 0], α=0)

Calculate ψ and q in a layered QG model using coefficients and vortex parameters

Arguments:

a: M x N array of coefficients, Array

(U, ℓ): vortex speed and radius, Numbers

(R, β): Rossby radii and (y) PV gradients in each layer, Numbers or Vectors

grid: grid structure containing x, y, and Krsq

x₀: position of vortex center, vector (default: [0, 0])

α: initial angle of vortex, Number (default: 0)

QGDipoles.Calc_ψb — Function

Function: Calc_ψb(a, U, ℓ, R, β, grid, x₀=[0, 0], α=0)

Calculate SQG fields ψ and b using coefficients and vortex parameters

Arguments:

a: M x 1 array of coefficients, Array

(U, ℓ): vortex speed and radius, Numbers

R: vector of [R, R'], Vector

β: beta-plane (y) PV gradient, Number

grid: grid structure containing x, y, and Krsq

x₀: position of vortex center, vector (default: [0, 0])

α: initial angle of vortex, Number (default: 0)

Note: Here R is the baroclinic Rossby radius, R = NH/f, and R' = R₀²/R where R₀ is the barotropic Rossby radius, R₀ = √(gH)/f. For infinite depth, R' = g/(fN).

QGDipoles.Calc_uv — Function

Function: Calc_uv(ψ, grid)

Calculate the velocity fields from ψ using (u, v) = (-∂ψ/∂y, ∂ψ/∂x)

Arguments:

ψ: streamfunction, Array

grid: grid structure containing kr and l

QGDipoles.ΔNCalc — Function

Function: ΔNCalc(K², R, β, U=1)

Defines the Δ_N(β) matrix used to invert for ψ and q

Arguments:

K²: value of k²+l² in Fourier space, Array

(R, β): Rossby radii and (y) PV gradients in each layer, Numbers or Vectors

U: vortex speed, Number (default: 1)

QGDipoles.CreateModonLQG — Function

Function: CreateModonLQG(grid, M, U=1, ℓ=1, R=1, β=0, ActiveLayers=1, x₀=[0, 0], α=0; K₀=Nothing, a₀=Nothing, tol=1e-6)

High level wrapper function for calculating ψ and q for the Layered QG model using given parameters

Arguments:

grid: grid structure containing x, y, and Krsq

M: number of coefficient to solve for, Integer (default: 8)

(U, ℓ): vortex speed and radius, Numbers (default: (1, 1))

(R, β): Rossby radii and (y) PV gradients in each layer, Numbers or Vectors, (default: (1, 0))

ActiveLayers: vector of 1s or 0s where 1 denotes an active layer, Number or Vector, (default: [1,..,1])

x₀: position of vortex center, vector (default: [0, 0])

α: initial angle of vortex, Number (default: 0)

K₀, a₀: initial guesses for K and a, Arrays or Nothings (default: Nothing)

tol: error tolerance passed to QuadGK and NLSolve functions, Number (default: 1e-6)

Note: provide values of K₀ and a₀ for active layers ONLY.

QGDipoles.CreateModonSQG — Function

Function: CreateModonSQG(grid, M, U=1, ℓ=1, R=[Inf, Inf], β=0, x₀=[0, 0], α=0; K₀=Nothing, a₀=Nothing, tol=1e-6)

High level wrapper function for calculating ψ and b for the SQG model using given parameters

Arguments:

grid: grid structure containing x, y, and Krsq

M: number of coefficient to solve for, Integer (default: 12)

(U, ℓ): vortex speed and radius, Numbers (default: (1, 1))

R: vector of [R, R'], Vector (default: [Inf, Inf])

β: beta-plane (y) PV gradient, Number (default: 0)

x₀: position of vortex center, vector (default: [0, 0])

α: initial angle of vortex, Number (default: 0)

K₀, a₀: initial guesses for K and a, Arrays or Nothings (default: Nothing)

tol: error tolerance passed to QuadGK and NLSolve functions, Number (default: 1e-6)

Note: Here R is the baroclinic Rossby radius, R = NH/f, and R' = R₀²/R where R₀ is the barotropic Rossby radius, R₀ = √(gH)/f. For infinite depth, R' = g/(fN).

QGDipoles.CreateLCD — Function

Function: CreateLCD(grid, U=1, ℓ=1, x₀=[0, 0], α=0)

High level wrapper function for calculating ψ and q for the Lamb-Chaplygin dipole using given parameters

Arguments:

grid: grid structure containing x, y, and Krsq

(U, ℓ): vortex speed and radius, Numbers (default: (1, 1))

x₀: position of vortex center, vector (default: [0, 0])

α: initial angle of vortex, Number (default: 0)

Note: This function uses the analytic solution for the LCD to calculate ψ and q.

QGDipoles.CreateLRD — Function

Function: CreateLRD(grid, U=1, ℓ=1, R=1, β=0, x₀=[0, 0], α=0)

High level wrapper function for calculating ψ and q for the Larichev-Reznik dipole using given parameters

Arguments:

grid: grid structure containing x, y, and Krsq

(U, ℓ): vortex speed and radius, Numbers (default: (1, 1))

(R, β): Rossby radii and (y) PV gradient, Numbers, (default: (1, 0))

x₀: position of vortex center, vector (default: [0, 0])

α: initial angle of vortex, Number (default: 0)

Note: This function uses the analytic solution for the LRD to calculate ψ and q.

QGDipoles.Eval_ψ_SQG — Function

Function: Eval_ψ_SQG(grid, ψ, z=[0], U=1, R=[Inf, Inf], β=0)

Evaluates ψ at specified depths, z in [-R, 0], for the SQG problem

Arguments:

grid: grid structure containing x, y, and Krsq

ψ: surface streamfunction, calculated using Calc_ψb or CreateModonSQG

z: vector of depths (default: [0])

U: vortex speed, Number (default: 1)

R: vector of [R, R'], Vector (default: [Inf, Inf])

β: beta-plane (y) PV gradient, Number (default: 0)

Note: Here R is the baroclinic Rossby radius, R = NH/f, and R' = R₀²/R where R₀ is the barotropic Rossby radius, R₀ = √(gH)/f. For infinite depth, R' = g/(fN).

QGDipoles.Eval_q_SQG — Function

Function: Eval_q_SQG(grid, ψ, z=[0], U=1, R=[Inf, Inf], β=0)

Evaluates q at specified depths, z in [-R, 0], for the SQG problem

Arguments:

grid: grid structure containing x, y, and Krsq

ψ: surface streamfunction, calculated using Calc_ψb or CreateModonSQG

z: vector of depths (default: [0])

U: vortex speed, Number (default: 1)

R: vector of [R, R'], Vector (default: [Inf, Inf])

β: beta-plane (y) PV gradient, Number (default: 0)

Note: Here R is the baroclinic Rossby radius, R = NH/f, and R' = R₀²/R where R₀ is the barotropic Rossby radius, R₀ = √(gH)/f. For infinite depth, R' = g/(fN).

QGDipoles.Eval_b_SQG — Function

Function: Eval_b_SQG(grid, ψ, z=[0], U=1, R=[Inf, Inf], β=0)

Evaluates b at specified depths, z in [-R, 0], for the SQG problem

Arguments:

grid: grid structure containing x, y, and Krsq

ψ: surface streamfunction, calculated using Calc_ψb or CreateModonSQG

z: vector of depths (default: [0])

U: vortex speed, Number (default: 1)

R: vector of [R, R'], Vector (default: [Inf, Inf])

β: beta-plane (y) PV gradient, Number (default: 0)

Note: Here R is the baroclinic Rossby radius, R = NH/f, and R' = R₀²/R where R₀ is the barotropic Rossby radius, R₀ = √(gH)/f. For infinite depth, R' = g/(fN).